[EGD-2592] bsp: change heap and code size
Change SDRAM configuration from 4M for code and 13M (which is invalid)
for heaps to 6M for code and 10M for heaps.

Usermem heap's size has been decreased to 9M.

Use heap start symbol __sdram_cached_start for MPU configuration instead
of using magic values which are easy to miss when they are need to be
changed.

This change is required to be able to fit in libphonenumber.