build: optimize third party in debug
To minimize size footprint and to be able to fit in the RT1051 target,
add optimization options in debug configuration for third party modules
which are directly added to the sources.

This behaviour is controlled by the THIRD_PARTY_DEBUG_OPTIMIZE boolean
option.

Add extra warning supressing option for the protobuf sources as its
build was starting to failing on newer compiler.

Signed-off-by: Marcin Smoczyński <smoczynski.marcin@gmail.com>