[EGD-3556] allow disabling applications
RT1051 SDRAM is not big enough to fit full Debug build. To be able to
debug on the RT1051 add options to choose precisely which applications
you want to include in the build allowing to decrease debug firmware
size.

Enable size optimization for applications (configurable) in debug mode
and allow to specify with cmake options which applications can be
optimized with.

Remove -Og debug flags which break debugging.

Signed-off-by: Marcin Smoczyński <smoczynski.marcin@gmail.com>