[EGD-3606] fix speed dial on contact merge
When there is a conflict on a speed dial user is promptod if he/she
wants to move speed dial number to the updated contact. Currently this
does not work because of invalid meta update in the DialogYesNoIconTxt.

Fix meta update in the above mentioned class to enable proper speed dial
moving to the updated contact.