[EGD-3714] rt1051: dtcm ld cleanup
There is a memory shortage in the DTCM region. In order to be able to
add new features linker script for the rt1051 needs to be cleanup.

A couple of issues has been fixed:
* double stack definition blocking 4kB
* excessive heap definition blocking 4kB
* a part of DTCM had not been used blocking 24kB.

Data has been split to user and system part, user data has been moved to
the SDRAM, therefore new data will be placed there. To place it inside
the MCU an explicit statement has to be used.

Added a basic description of memory layout and a couple of comments.

Signed-off-by: Marcin Smoczyński <smoczynski.marcin@gmail.com>