[EGD-4016] FreeRTOS fat as a submodule
Currently VFS class has a separate implementation
which uses native Linux filesystem, so currently the
filesystem can behave differ on the emulator than in the phone.
In the linux emulator VFS should use same FAT code as the phone
and emulate block device as a file on the linux platform.