[EGD-4544] doc: update requirements for commits
Change the requirements for commits' description.

In order to be able to remove the obligation to change the project
changelog with every PR, provide requirements for commits' title and
description. The first sentence of each commit will be included in
the changelog directly, while the rest of it should help understand
others motivation and result of a change.