[EGD-6855] Add no modem notification on tethering
This commit provides the implementation of functionality that blocks
sms/calls notifications on home screen when tethering is active.
Previously, the notifications could be visible if the notifications
were visible prior to tethering activation.

By the design, the functionality should not clear the notifications,
just temporary hide them. That is achieved by implementation ot the
functionality in NotificationModel which is an UI presenter.