[BH-913] BGSounds audio implementation
The following commit provides an implemention of audio
features for the BGSounds application, that is:
* playback of selected record in either looped or
singleshot mode
* pausing, resuming and stopping the playback

The implementation does not cover a sinchronization
of displayed time with a acctual state of playback in
singleshot mode and changes of a playback volume. The
features will be addressed in a subsuquent tasks.