[CP-599] Add crc hash to the last chunk of sent file
CRC was counted in one go before starting to send
chunks of the file, which, in the case of a large file,
caused communication to stop for a long while.
In the current implementation, CRC is counted
incrementally while sending individual file chunks.
It is then appended to the last chunk of the file.