[BH-1080] Fix Bgsounds timeout and volume
The following commit provides a fix for:
 * BH-1080: bug appeard only in case off timer's
auto turn off set to off.  THe OnStopCallback
passed to async stop request was not processed due
to recieving Failed RetCode, which is expected in
singleShot mode. Since there is no need for calling
BGSoundsPlayer::stop on timeout in singleShot mode,
the check was transfered to presenter's onFinished
method.
 * BH-1079: user could set Volume to 0. System-wdie
min volume is 0 so there was presenter-level check added
to restrict minimal Volume value to 1.