[EGD-7400] Device won't reconnect to MC when unplugged
In case of disconnection from MC during reception of large messages
(e.g. to FS endpoint), the parser may be left in a state expecting
further parts of the message. This causes incorrect handling of new
messages arriving after reconnection to MC.
Added a timer to reset a Parser when transfer times out.