[EGD-8208] Removed unused ping, added simple Async
Removed unused code: ping and stale messages
Detached UnicastSync send and sync part so that we would be able
to await for message in the code.
Added sys::Async to be able to handle situations where we want to
sync to the response, but after some time.
Fixed messages logging: for some reason demangling abi use caused
bad system behaviour