[BH-1519] Shared alarm model instance
Alarm model was moved to the bell-specific
application class. By doing this, each application will
have access to its unique and private instance of the model.
Later, it can be propagated to the specific presenters etc.
The next step should be to use only the one instance
of the alarm model and propagate it to the each application.