[BH-1554] Add delay after init external oscillator
Some Harmony devices are struggling to switch between
internal to external oscillators. According to the newest
version of library (lpm.c - Low Power Mode) the NXP
is recommending to add at least 200us delay after
enabling the external oscillator.

Their implementation is not full beause the speed of
the CPU used in the delay function is hardcoded.
So if the CPU has frequency 4MHz the delay takes ~26ms.

To avoid that situation we use the proper speed of the CPU
in the delay function.