[BH-1628] Reinit I2C after bus error
A lack of timeout in I2C could cause
the worker could get stuck in a loop
if the communication has been broken somehow.
The Harmony is unresponsive in this case.
Only solution is just perform hardware reset using
Frontlight and Back buttons simultaneously.

There is also additional 5 attemps to read and write
transactions if the bus error occures in Fuel Gauge driver.
It should be sufficient but if it would be still persist the system
just will read 0% SOC and shutdown system.