[MOS-663] Regression fix
Fixed a regression that the SMS template list showed always
the state representing no templates defined, regardless if
there were any in the database.

There's a minor side effect that querying the database and
refreshing the window is doubled which can be seen in
debugging, but it's unnoticeable in usage.