[MOS-973] Fix for a ghost call after quick click back key to end a call
Fixed a scenario where the user click to make a call to some contact and
then quickly click BACK (right function button) to end the call and despite
that Pure was still calling without showing any information.
This fix prevent to handle key press event as a key longpress even after
application lose a focus, or change a state.