[BH-1705] Fixed back button power off timer to 10s
When holding the back button, the device would prompt to
power off after 5s. This was not inline with the manual. Now
the prompt will come up after 10s which is correct by the
manual. In the code, the actual time is set to 9s because
an additional 1s delay is caused by the OS and screen
refresh rate.