[BH-1705] Adjusted power off timer to >=10s
When holding the back button, the device would prompt to
power off after ~9-10s. This was not inline with the manual.
Now the prompt will come up just after 10s. Previously the
timer was set to 9s to account for delays caused be the OS
and display, however it is now decided that the only thing
that should matter is the time you actually have to hold the
button for the prompt to show. This time is more or less exactly
the same as the time set in the code.