[BH-1739] Fix problem with charging the device
For some chargers, the USB can't detect plug events and returns an error code.
Maybe it is a hardware issue or the charger is not USB-compliant. However, we can try to check the voltage.
If the voltage is valid it means that the charger is connected and we can enable charging.
Also changed the name of the charging driver and did a small refactor.