[MOS-195] Phone number input style unification
Unification of input style for phone number. Now every time when the user will provide
a phone number in gui the right function key (RF key) will be work as CLEAR.
This means that if any number is in the phone number field, the RF key will remove
the last character. Otherwise it'll work as typical BACK button.
This changes also provide ability to add * and # sign to number.