[BH-1817] Fix alarm when the onboarding is active
The alarm shouldn't activate itself when
the onboarding is in progress. So now we check
the onboarding status before checking alarms.

Increased size of the onboarding application stack.
This should resolve the issue when somehow, the alarm is set
but there is no stack that causes the application to crash.