[BH-1822] Disable Address Sanitizer
Due to using wrappers for the glibc syscalls the ASAN
doesn't work well. We also had trouble with integrating this
tool with GCC12.
The ASAN is still available but is turned off by default.
Use -DLINUX_ENABLE_SANITIZER=ON to turn on.
Our recommendation is to use Valgrind as the runtime
profiling tool. It's added for visual studio code.