[BH-2013] Add label with the name of the application on the countdown screen
Harmony's UI will be consistent so that on each screen showing the
elapsed time on a timer (Meditation, Relaxation and Focus Timer, Power
nap) will have words indicating which feature is active. This is how it
is done with the focus timer and has been added to other timers as well.