[BH-2026] Add mechanism for getting previously OS version
Task relates to the What’s new feature that needs to know which version of MuditaOS was installed before the update.
Normally What’s new uses a database to get the previous  version, but after the first run this field will be empty.
To bypass this issue the PureRecovery needs to save the previous version in recovery_status.json that is generated after the update process.